before merge fran

+++ client/src/redux/actions.js
@@ -96,7 +96,7 @@ export const favorite = (course) => (dispatch) => {
 
 export const getFavorites = (id) => async (dispatch) => {
   const user = await axios.get(`http://localhost:3001/api/users/${id}`);
-  console.log(user);
+  // console.log(user);
   dispatch(getFavoriteCourse(user.data.favorites));
 };
 
@@ -244,22 +244,55 @@ export const googleUserLogin = (payload) => async (dispatch) => {
   const response = await axios
     .post("http://localhost:3001/api/users/google_login", payload)
     .then((res) => {
-      dispatch(getSession(res.data));
-      window.localStorage.setItem("user", JSON.stringify(res.data));
+      if (res.data.user.status === "pending") {
+        Swal.fire({
+          title: "Ups Something Happens",
+          // text: "Can't create video please try again",
+          text: "Pending Account. Please Verify Your Email!",
+          icon: "error",
+          confirmButtonText: "OK",
+        });
+      } else {
+        Swal.fire({
+          title: "Successful login",
+          text: "You are being redirected to the home",
+          icon: "success",
+          timer: 2000,
+          showConfirmButton: false,
+        });
+        dispatch(getSession(res.data));
+        window.localStorage.setItem("user", JSON.stringify(res.data));
+      }
     })
-    .catch((e) => console.log(e));
-
+    .catch((error) =>
+      Swal.fire({
+        title: "Ups Something Happens",
+        // text: "Can't create video please try again",
+        text: error.response.data.error,
+        icon: "error",
+        confirmButtonText: "OK",
+      })
+    );
   return response;
 };
-export const gitHubLogin = (payload) => async (dispatch) => {
+export const gitHubLogin = () => async (dispatch) => {
   const response = await axios
     .get("http://localhost:3001/api/auth/github_login")
     .then((res) => {
+      console.log(res.data);
       dispatch(getSession(res.data));
       window.localStorage.setItem("user", JSON.stringify(res.data));
     })
-    .catch((e) => console.log(e));
-
+    .catch((error) => {
+      console.log(error);
+      Swal.fire({
+        title: "Ups Something Happens",
+        // text: "Can't create video please try again",
+        text: error.response.data.error,
+        icon: "error",
+        confirmButtonText: "OK",
+      });
+    });
   return response;
 };
